fix: deduplicate the audience-member job and follow the repo's job conventions
Addresses the pre-merge review of this branch.

The delay alone collapsed nothing: a client verifying the same license
repeatedly enqueued one job per call, and each one redid the same full
rewrite of the same audience_members row, which relocated the pile-up into
Sidekiq rather than removing it. Adding `lock: :until_executed` makes the
deduplication real — verified by enqueuing five times for one purchase and
seeing a single job. Deduplicating on purchase_id is safe because the job
derives everything it writes from current database state.

Also brings the job in line with the contributing guide: `Job` suffix rather
than the legacy `Worker` one, and the `low` queue, since the guide reserves
`default` for work that is time-sensitive and this is read only by email
audience filters.

Finally, `find` replaces `find_by` plus a silent return. Purchases are
soft-deleted, so a genuinely missing row means a lagging replica read, and
the ElasticsearchIndexerWorker this change mirrors raises in exactly that
case. Swallowing it would drop the audience update with no signal; raising
hands it to the retries. The spec now pins the raise, and covers a purchase
that no longer qualifies for the audience.
